refactor(factoryconfig): split metadata and layout mapping
