+++ pkg/transports/mapping/factoryconfig/factory_config_mapping.go
@@ -1,12 +1,7 @@
-// backendsizecheck:ignore-file canonical public factory contract mapping remains consolidated until dedicated mapper seams are extracted from pkg/config.
-// pkgmaintcheck:ignore-file-lines canonical public factory contract mapping remains consolidated until dedicated mapper seams are extracted from pkg/config.
+// Factory mapping helpers for the remaining work-type, resource, worker, workstation, and compatibility projections.
 package factoryconfig
 
 import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"io"
 	"sort"
 	"strings"
 
@@ -16,864 +11,8 @@ import (
 
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	workercompatibility "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/mapping/factorydefinition/retiredboundary"
 	contentcontract "github.com/portpowered/infinite-you/pkg/transports/mapping/workcontent"
 )
-
-// FactoryConfigMapper maps between on-disk factory configuration payloads and
-// canonical in-memory config structures.
-type FactoryConfigMapper struct{}
-
-// NewFactoryConfigMapper returns the canonical mapper used across config loading
-// and serialization paths.
-func NewFactoryConfigMapper() *FactoryConfigMapper {
-	return &FactoryConfigMapper{}
-}
-
-type generatedFactoryBoundary struct {
-	generated factoryapi.Factory
-}
-
-const generatedFactoryBoundaryErrorPrefix = "decode factory generated-schema boundary"
-
-// Expand parses and normalizes a user-provided factory payload into the internal
-// canonical configuration representation.
-func (m *FactoryConfigMapper) Expand(data []byte) (*interfaces.FactoryConfig, error) {
-	boundary, err := decodeGeneratedFactoryBoundaryJSON(data)
-	if err != nil {
-		return nil, err
-	}
-
-	cfg, err := FactoryConfigFromOpenAPI(boundary.generated)
-	if err != nil {
-		return nil, err
-	}
-	return &cfg, nil
-}
-
-func decodeGeneratedFactoryBoundaryJSON(data []byte) (generatedFactoryBoundary, error) {
-	if err := retiredboundary.RejectGeneratedBoundaryAliases(data); err != nil {
-		return generatedFactoryBoundary{}, fmt.Errorf("%s: %w", generatedFactoryBoundaryErrorPrefix, err)
-	}
-	normalizedData, err := normalizeFactoryInputJSON(data)
-	if err != nil {
-		return generatedFactoryBoundary{}, fmt.Errorf("%s: %w", generatedFactoryBoundaryErrorPrefix, err)
-	}
-	if err := retiredboundary.RejectExhaustionRulesField(normalizedData); err != nil {
-		return generatedFactoryBoundary{}, fmt.Errorf("%s: %w", generatedFactoryBoundaryErrorPrefix, err)
-	}
-	if err := retiredboundary.RejectFanInField(normalizedData); err != nil {
-		return generatedFactoryBoundary{}, fmt.Errorf("%s: %w", generatedFactoryBoundaryErrorPrefix, err)
-	}
-	if err := retiredboundary.RejectCronIntervalField(normalizedData); err != nil {
-		return generatedFactoryBoundary{}, fmt.Errorf("%s: %w", generatedFactoryBoundaryErrorPrefix, err)
-	}
-	if err := validatePortableLayoutBoundaryJSON(normalizedData); err != nil {
-		return generatedFactoryBoundary{}, fmt.Errorf("%s: %w", generatedFactoryBoundaryErrorPrefix, err)
-	}
-
-	apiCfg, err := decodeGeneratedFactoryBoundary(normalizedData)
-	if err != nil {
-		return generatedFactoryBoundary{}, fmt.Errorf("%s: %w", generatedFactoryBoundaryErrorPrefix, err)
-	}
-	return generatedFactoryBoundary{
-		generated: apiCfg,
-	}, nil
-}
-
-func decodeGeneratedFactoryBoundary(data []byte) (factoryapi.Factory, error) {
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.DisallowUnknownFields()
-
-	var apiCfg factoryapi.Factory
-	if err := decoder.Decode(&apiCfg); err != nil {
-		return factoryapi.Factory{}, fmt.Errorf("unmarshal factory api model: %w", err)
-	}
-	if err := ensureFactoryBoundaryEOF(decoder); err != nil {
-		return factoryapi.Factory{}, err
-	}
-	if err := validateGeneratedFactoryBoundary(apiCfg); err != nil {
-		return factoryapi.Factory{}, err
-	}
-	return apiCfg, nil
-}
-
-func ensureFactoryBoundaryEOF(decoder *json.Decoder) error {
-	var extra json.RawMessage
-	if err := decoder.Decode(&extra); err != nil {
-		if err == io.EOF {
-			return nil
-		}
-		return fmt.Errorf("unmarshal factory api model: %w", err)
-	}
-	return fmt.Errorf("unmarshal factory api model: unexpected trailing JSON value")
-}
-
-func validateGeneratedFactoryBoundary(apiCfg factoryapi.Factory) error {
-	if strings.TrimSpace(string(apiCfg.Name)) == "" {
-		return fmt.Errorf("factory.name is required")
-	}
-	if apiCfg.Examples != nil {
-		for index, example := range *apiCfg.Examples {
-			if strings.TrimSpace(example.Name) == "" {
-				return fmt.Errorf("factory.examples[%d].name must be a non-empty string", index)
-			}
-			if example.Args == nil {
-				return fmt.Errorf("factory.examples[%d].args is required", index)
-			}
-		}
-	}
-	return nil
-}
-
-// Flatten serializes an internal factory configuration into canonical JSON that is
-// stable for persisted output and downstream tooling.
-func (m *FactoryConfigMapper) Flatten(cfg *interfaces.FactoryConfig) ([]byte, error) {
-	apiCfg, err := factoryAPIFromInternalConfig(cfg)
-	if err != nil {
-		return nil, err
-	}
-	if isDefaultPetriOrchestratorAPI(apiCfg.Orchestrator) {
-		apiCfg.Orchestrator = nil
-	}
-
-	raw, err := json.Marshal(apiCfg)
-	if err != nil {
-		return nil, fmt.Errorf("marshal factory api model: %w", err)
-	}
-
-	var decoded any
-	if err := json.Unmarshal(raw, &decoded); err != nil {
-		return nil, fmt.Errorf("decode factory api payload: %w", err)
-	}
-	canonical := normalizeFactoryOutputJSONKeys(decoded)
-	dropSupportedPortableBundledInlineContent(canonical)
-	encoded, err := json.Marshal(canonical)
-	if err != nil {
-		return nil, fmt.Errorf("normalize factory config keys: %w", err)
-	}
-	return encoded, nil
-}
-
-func factoryAPIFromInternalConfig(cfg *interfaces.FactoryConfig) (factoryapi.Factory, error) {
-	if cfg == nil {
-		return factoryapi.Factory{}, nil
-	}
-	if err := validateInternalFactoryMetadata(cfg); err != nil {
-		return factoryapi.Factory{}, err
-	}
-	examples, err := invocationExamplesAPIFromInternal(cfg.Examples)
-	if err != nil {
-		return factoryapi.Factory{}, err
-	}
-
-	return factoryapi.Factory{
-		Name:                factoryReferenceName(cfg),
-		Description:         NameValueAPIFromInternal(cfg.Description),
-		Id:                  stringPtrIfNotEmpty(cfg.Project),
-		Version:             hybridLogicalTimestampPtr(cfg.Version),
-		Guards:              factoryGuardsAPIFromInternal(cfg.Guards),
-		InputTypes:          inputTypesAPIFromInternal(cfg.InputTypes),
-		InvocationReturn:    invocationReturnAPIFromInternal(cfg.InvocationReturn),
-		InvocationSignature: invocationSignatureAPIFromInternal(cfg.InvocationSignature),
-		Examples:            examples,
-		Webhooks:            factoryWebhooksAPIFromInternal(cfg.Webhooks),
-		Orchestrator:        orchestratorAPIFromInternal(cfg),
-		WorkTypes:           workTypesAPIFromInternal(cfg.WorkTypes),
-		Resources:           resourcesAPIFromInternal(cfg.Resources),
-		SupportingFiles:     resourceManifestAPIFromInternal(cfg.ResourceManifest),
-		Layout:              factoryLayoutAPIFromInternal(cfg.Layout),
-		Workers:             workersAPIFromInternal(cfg.Workers, cfg.Workstations),
-		Workstations:        workstationsAPIFromInternal(cfg.Workstations, workerTypesByName(cfg.Workers)),
-	}, nil
-}
-
-func factoryWebhooksAPIFromInternal(values []interfaces.FactoryWebhookConfig) *[]factoryapi.FactoryWebhook {
-	if len(values) == 0 {
-		return nil
-	}
-	result := make([]factoryapi.FactoryWebhook, len(values))
-	for index, value := range values {
-		filter := factoryapi.FactoryWebhookFilter{
-			EventTypes: make([]factoryapi.FactoryWebhookEventType, len(value.Filter.EventTypes)),
-		}
-		for eventIndex, eventType := range value.Filter.EventTypes {
-			filter.EventTypes[eventIndex] = factoryapi.FactoryWebhookEventType(eventType)
-		}
-		if len(value.Filter.DispatchStatuses) > 0 {
-			statuses := make([]factoryapi.FactoryWebhookDispatchStatus, len(value.Filter.DispatchStatuses))
-			for statusIndex, status := range value.Filter.DispatchStatuses {
-				statuses[statusIndex] = factoryapi.FactoryWebhookDispatchStatus(status)
-			}
-			filter.DispatchStatuses = &statuses
-		}
-		result[index] = factoryapi.FactoryWebhook{
-			Name:             value.Name,
-			Enabled:          value.Enabled,
-			Url:              value.URL,
-			SigningSecretRef: value.SigningSecretRef,
-			Filter:           filter,
-			DeliveryPolicy:   factoryWebhookDeliveryPolicyAPIFromInternal(value.DeliveryPolicy),
-		}
-	}
-	return &result
-}
-
-func factoryWebhookDeliveryPolicyAPIFromInternal(value *interfaces.FactoryWebhookDeliveryPolicyConfig) *factoryapi.FactoryWebhookDeliveryPolicy {
-	if value == nil {
-		return nil
-	}
-	result := &factoryapi.FactoryWebhookDeliveryPolicy{
-		RequestTimeout: value.RequestTimeout,
-		InitialBackoff: value.InitialBackoff,
-		MaxBackoff:     value.MaxBackoff,
-	}
-	if value.MaxAttempts != nil {
-		result.MaxAttempts = value.MaxAttempts
-	}
-	if value.BackoffMultiplier != nil {
-		multiplier := float32(*value.BackoffMultiplier)
-		result.BackoffMultiplier = &multiplier
-	}
-	return result
-}
-
-func validateInternalFactoryMetadata(cfg *interfaces.FactoryConfig) error {
-	if err := validateInternalNameValue(cfg.Description, "factory.description"); err != nil {
-		return err
-	}
-	for index := range cfg.WorkTypes {
-		if err := validateInternalNameValue(cfg.WorkTypes[index].Description, fmt.Sprintf("factory.workTypes[%d].description", index)); err != nil {
-			return err
-		}
-	}
-	for index := range cfg.Workers {
-		if err := validateInternalNameValue(cfg.Workers[index].Description, fmt.Sprintf("factory.workers[%d].description", index)); err != nil {
-			return err
-		}
-	}
-	for index := range cfg.Workstations {
-		if err := validateInternalNameValue(cfg.Workstations[index].Description, fmt.Sprintf("factory.workstations[%d].description", index)); err != nil {
-			return err
-		}
-	}
-	for index := range cfg.Examples {
-		if strings.TrimSpace(cfg.Examples[index].Name) == "" {
-			return fmt.Errorf("factory.examples[%d].name must be a non-empty string", index)
-		}
-		if err := validateInternalNameValue(&cfg.Examples[index].Description, fmt.Sprintf("factory.examples[%d].description", index)); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateInternalNameValue(value *interfaces.NameValueConfig, path string) error {
-	if value == nil {
-		return nil
-	}
-	if err := value.Validate(); err != nil {
-		return fmt.Errorf("%s.%w", path, err)
-	}
-	return nil
-}
-
-// NameValueAPIFromInternal maps a validated internal localized value into the
-// generated public contract without sharing mutable collection storage.
-func NameValueAPIFromInternal(value *interfaces.NameValueConfig) *factoryapi.NameValue {
-	if value == nil {
-		return nil
-	}
-	result := &factoryapi.NameValue{
-		Type:  factoryapi.NameValueType(value.Type),
-		Value: value.Value,
-		Id:    stringPtrIfNotEmpty(value.ID),
-	}
-	if len(value.Locales) > 0 {
-		locales := append([]string(nil), value.Locales...)
-		result.Locales = &locales
-	}
-	if len(value.Values) > 0 {
-		values := make(map[string]string, len(value.Values))
-		for locale, localized := range value.Values {
-			values[locale] = localized
-		}
-		result.Values = &values
-	}
-	return result
-}
-
-func invocationSignatureAPIFromInternal(value *interfaces.InvocationSignatureConfig) *factoryapi.FactoryInvocationSignature {
-	if value == nil {
-		return nil
-	}
-	return &factoryapi.FactoryInvocationSignature{
-		Parameters:                 invocationParametersAPIFromInternal(value.Parameters),
-		UnknownNamedArgumentPolicy: invocationUnknownNamedArgumentPolicyPtr(value.UnknownNamedArgumentPolicy),
-		OutputContract:             invocationOutputContractAPIFromInternal(value.OutputContract),
-	}
-}
-
-func invocationParametersAPIFromInternal(parameters []interfaces.InvocationParameterConfig) *[]factoryapi.FactoryInvocationParameter {
-	if len(parameters) == 0 {
-		return nil
-	}
-	values := make([]factoryapi.FactoryInvocationParameter, len(parameters))
-	for i, parameter := range parameters {
-		values[i] = factoryapi.FactoryInvocationParameter{
-			Name:          parameter.Name,
-			Description:   stringPtrIfNotEmpty(parameter.Description),
-			ExternalName:  stringPtrIfNotEmpty(parameter.ExternalName),
-			Aliases:       stringSlicePtr(parameter.Aliases),
-			TypeHint:      invocationParameterTypeHintPtr(parameter.TypeHint),
-			ValueMode:     invocationParameterValueModePtr(parameter.ValueMode),
-			Required:      boolPtrIfTrue(parameter.Required),
-			Sensitive:     boolPtrIfTrue(parameter.Sensitive),
-			Choices:       stringSlicePtr(parameter.Choices),
-			DefaultValue:  stringPtrIfNotEmpty(parameter.DefaultValue),
-			DefaultValues: stringSlicePtr(parameter.DefaultValues),
-			Bindings:      invocationParameterBindingsAPIFromInternal(parameter.Bindings),
-		}
-	}
-	return &values
-}
-
-func invocationParameterBindingsAPIFromInternal(bindings []interfaces.InvocationParameterBindingConfig) *[]factoryapi.FactoryInvocationParameterBinding {
-	if len(bindings) == 0 {
-		return nil
-	}
-	values := make([]factoryapi.FactoryInvocationParameterBinding, len(bindings))
-	for i, binding := range bindings {
-		values[i] = factoryapi.FactoryInvocationParameterBinding{
-			Kind:     factoryapi.FactoryInvocationParameterBindingKind(binding.Kind),
-			Position: intPtrIfNonZero(binding.Position),
-		}
-	}
-	return &values
-}
-
-func invocationUnknownNamedArgumentPolicyPtr(value string) *factoryapi.FactoryInvocationUnknownNamedArgumentPolicy {
-	if value == "" {
-		return nil
-	}
-	policy := factoryapi.FactoryInvocationUnknownNamedArgumentPolicy(value)
-	return &policy
-}
-
-func invocationParameterTypeHintPtr(value string) *factoryapi.FactoryInvocationParameterTypeHint {
-	if value == "" {
-		return nil
-	}
-	typeHint := factoryapi.FactoryInvocationParameterTypeHint(value)
-	return &typeHint
-}
-
-func invocationParameterValueModePtr(value string) *factoryapi.FactoryInvocationParameterValueMode {
-	if value == "" {
-		return nil
-	}
-	valueMode := factoryapi.FactoryInvocationParameterValueMode(value)
-	return &valueMode
-}
-
-func invocationOutputContractAPIFromInternal(value *interfaces.InvocationOutputContractConfig) *factoryapi.FactoryInvocationOutputContract {
-	if value == nil {
-		return nil
-	}
-	return &factoryapi.FactoryInvocationOutputContract{
-		Mode:          invocationOutputContractModePtr(value.Mode),
-		PathParameter: stringPtrIfNotEmpty(value.PathParameter),
-		ContentType:   stringPtrIfNotEmpty(value.ContentType),
-		FileExtension: stringPtrIfNotEmpty(value.FileExtension),
-		Description:   stringPtrIfNotEmpty(value.Description),
-	}
-}
-
-func invocationOutputContractModePtr(value string) *factoryapi.FactoryInvocationOutputContractMode {
-	if value == "" {
-		return nil
-	}
-	mode := factoryapi.FactoryInvocationOutputContractMode(value)
-	return &mode
-}
-
-func invocationExamplesAPIFromInternal(examples []interfaces.InvocationExampleConfig) (*[]factoryapi.FactoryInvocationExample, error) {
-	if len(examples) == 0 {
-		return nil, nil
-	}
-	values := make([]factoryapi.FactoryInvocationExample, len(examples))
-	for i, example := range examples {
-		args, err := invocationExampleArgsAPIFromInternal(example.Args, fmt.Sprintf("factory.examples[%d].args", i))
-		if err != nil {
-			return nil, err
-		}
-		values[i] = factoryapi.FactoryInvocationExample{
-			Name:        example.Name,
-			Description: *NameValueAPIFromInternal(&example.Description),
-			Args:        args,
-		}
-	}
-	return &values, nil
-}
-
-func invocationExampleArgsAPIFromInternal(args map[string]interface{}, path string) (factoryapi.FactoryInvocationArguments, error) {
-	values := make(factoryapi.FactoryInvocationArguments, len(args))
-	for name, value := range args {
-		var union factoryapi.FactoryInvocationArguments_AdditionalProperties
-		switch typed := value.(type) {
-		case string:
-			if err := union.FromFactoryInvocationArguments0(typed); err != nil {
-				return nil, fmt.Errorf("%s.%s: %w", path, name, err)
-			}
-		case []string:
-			if err := union.FromFactoryInvocationArguments1(append([]string(nil), typed...)); err != nil {
-				return nil, fmt.Errorf("%s.%s: %w", path, name, err)
-			}
-		default:
-			return nil, fmt.Errorf("%s.%s must be a string or array of strings", path, name)
-		}
-		values[name] = union
-	}
-	return values, nil
-}
-
-func validatePortableLayoutBoundaryJSON(data []byte) error {
-	var root map[string]any
-	if err := json.Unmarshal(data, &root); err != nil {
-		return fmt.Errorf("decode layout validation payload: %w", err)
-	}
-
-	layoutValue, ok := root["layout"]
-	if !ok || layoutValue == nil {
-		return nil
-	}
-
-	layout, ok := layoutValue.(map[string]any)
-	if !ok {
-		return fmt.Errorf("layout must be an object")
-	}
-	if err := requireNumber(layout, "schemaVersion", "layout"); err != nil {
-		return err
-	}
-	totalImageBytes := 0
-	if err := validateLayoutNodeArray(layout, "nodes", "layout", &totalImageBytes); err != nil {
-		return err
-	}
-	if err := validateLayoutEdgeArray(layout, "edges", "layout"); err != nil {
-		return err
-	}
-	if err := validateLayoutGroupArray(layout, "groups", "layout"); err != nil {
-		return err
-	}
-	if err := validateLayoutAnnotationArray(layout, "annotations", "layout", &totalImageBytes); err != nil {
-		return err
-	}
-	if err := validateOptionalPointObject(layout, "viewport", "layout", true, "zoom"); err != nil {
-		return err
-	}
-	if err := validateLayoutPreferences(layout, "preferences", "layout"); err != nil {
-		return err
-	}
-	return nil
-}
-
-func validateLayoutNodeArray(parent map[string]any, key string, path string, totalImageBytes *int) error {
-	values, ok, err := optionalObjectArray(parent, key, path)
-	if !ok || err != nil {
-		return err
-	}
-	seenEmptyStateNodeIDs := make(map[string]struct{}, len(values))
-	for index, node := range values {
-		nodePath := fmt.Sprintf("%s.%s[%d]", path, key, index)
-		if err := requireNonBlankString(node, "id", nodePath); err != nil {
-			return err
-		}
-		if err := validateOptionalPointObject(node, "position", nodePath, true); err != nil {
-			return err
-		}
-		if err := validateOptionalSizeObject(node, "size", nodePath, false); err != nil {
-			return err
-		}
-		imageBytes, err := validateLayoutNodeEmptyState(node, nodePath)
-		if err != nil {
-			return err
-		}
-		if imageBytes > 0 || nodeHasTextEmptyState(node) {
-			nodeID := node["id"].(string)
-			if _, duplicate := seenEmptyStateNodeIDs[nodeID]; duplicate {
-				return fmt.Errorf("%s.emptyState duplicates an empty state for canonical node %q", nodePath, nodeID)
-			}
-			seenEmptyStateNodeIDs[nodeID] = struct{}{}
-		}
-		*totalImageBytes += imageBytes
-		if *totalImageBytes > factoryLayoutEmbeddedImageTotalMaxBytes {
-			return fmt.Errorf("%s.emptyState.image.source.data exceeds the %d-byte Factory embedded-image budget", nodePath, factoryLayoutEmbeddedImageTotalMaxBytes)
-		}
-	}
-	return nil
-}
-
-func validateLayoutEdgeArray(parent map[string]any, key string, path string) error {
-	values, ok, err := optionalObjectArray(parent, key, path)
-	if !ok || err != nil {
-		return err
-	}
-	for index, edge := range values {
-		edgePath := fmt.Sprintf("%s.%s[%d]", path, key, index)
-		if err := requireString(edge, "id", edgePath); err != nil {
-			return err
-		}
-		if err := validateOptionalPointArray(edge, "waypoints", edgePath); err != nil {
-			return err
-		}
-		if err := validateOptionalPointObject(edge, "labelPosition", edgePath, false); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateLayoutGroupArray(parent map[string]any, key string, path string) error {
-	values, ok, err := optionalObjectArray(parent, key, path)
-	if !ok || err != nil {
-		return err
-	}
-	for index, group := range values {
-		groupPath := fmt.Sprintf("%s.%s[%d]", path, key, index)
-		if err := requireString(group, "id", groupPath); err != nil {
-			return err
-		}
-		if err := validateOptionalBoundsObject(group, "bounds", groupPath, true); err != nil {
-			return err
-		}
-		if err := requireStringArray(group, "nodeIds", groupPath); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateLayoutPreferences(parent map[string]any, key string, path string) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		return nil
-	}
-	record, ok := value.(map[string]any)
-	if !ok {
-		return fmt.Errorf("%s.%s must be an object", path, key)
-	}
-	direction, ok := record["direction"]
-	if !ok || direction == nil {
-		return nil
-	}
-	directionValue, ok := direction.(string)
-	if !ok {
-		return fmt.Errorf("%s.%s.direction must be a string", path, key)
-	}
-	switch directionValue {
-	case "UP", "DOWN", "LEFT", "RIGHT":
-		return nil
-	default:
-		return fmt.Errorf("%s.%s.direction must be one of UP, DOWN, LEFT, RIGHT", path, key)
-	}
-}
-
-func validateOptionalPointArray(parent map[string]any, key string, path string) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		return nil
-	}
-	items, ok := value.([]any)
-	if !ok {
-		return fmt.Errorf("%s.%s must be an array", path, key)
-	}
-	for index, item := range items {
-		record, ok := item.(map[string]any)
-		if !ok {
-			return fmt.Errorf("%s.%s[%d] must be an object", path, key, index)
-		}
-		itemPath := fmt.Sprintf("%s.%s[%d]", path, key, index)
-		if err := requireNumber(record, "x", itemPath); err != nil {
-			return err
-		}
-		if err := requireNumber(record, "y", itemPath); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateOptionalPointObject(parent map[string]any, key string, path string, required bool, extraRequiredKeys ...string) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		if required {
-			return fmt.Errorf("%s.%s is required", path, key)
-		}
-		return nil
-	}
-	record, ok := value.(map[string]any)
-	if !ok {
-		return fmt.Errorf("%s.%s must be an object", path, key)
-	}
-	pointPath := fmt.Sprintf("%s.%s", path, key)
-	if err := requireNumber(record, "x", pointPath); err != nil {
-		return err
-	}
-	if err := requireNumber(record, "y", pointPath); err != nil {
-		return err
-	}
-	for _, extraKey := range extraRequiredKeys {
-		if err := requireNumber(record, extraKey, pointPath); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateOptionalSizeObject(parent map[string]any, key string, path string, required bool) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		if required {
-			return fmt.Errorf("%s.%s is required", path, key)
-		}
-		return nil
-	}
-	record, ok := value.(map[string]any)
-	if !ok {
-		return fmt.Errorf("%s.%s must be an object", path, key)
-	}
-	sizePath := fmt.Sprintf("%s.%s", path, key)
-	if err := requireNumber(record, "width", sizePath); err != nil {
-		return err
-	}
-	if err := requireNumber(record, "height", sizePath); err != nil {
-		return err
-	}
-	return nil
-}
-
-func validateOptionalBoundsObject(parent map[string]any, key string, path string, required bool) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		if required {
-			return fmt.Errorf("%s.%s is required", path, key)
-		}
-		return nil
-	}
-	record, ok := value.(map[string]any)
-	if !ok {
-		return fmt.Errorf("%s.%s must be an object", path, key)
-	}
-	boundsPath := fmt.Sprintf("%s.%s", path, key)
-	if err := requireNumber(record, "x", boundsPath); err != nil {
-		return err
-	}
-	if err := requireNumber(record, "y", boundsPath); err != nil {
-		return err
-	}
-	if err := requireNumber(record, "width", boundsPath); err != nil {
-		return err
-	}
-	if err := requireNumber(record, "height", boundsPath); err != nil {
-		return err
-	}
-	return nil
-}
-
-func optionalObjectArray(parent map[string]any, key string, path string) ([]map[string]any, bool, error) {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		return nil, false, nil
-	}
-	items, ok := value.([]any)
-	if !ok {
-		return nil, true, fmt.Errorf("%s.%s must be an array", path, key)
-	}
-	result := make([]map[string]any, len(items))
-	for index, item := range items {
-		record, ok := item.(map[string]any)
-		if !ok {
-			return nil, true, fmt.Errorf("%s.%s[%d] must be an object", path, key, index)
-		}
-		result[index] = record
-	}
-	return result, true, nil
-}
-
-func requireString(parent map[string]any, key string, path string) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		return fmt.Errorf("%s.%s is required", path, key)
-	}
-	if _, ok := value.(string); !ok {
-		return fmt.Errorf("%s.%s must be a string", path, key)
-	}
-	return nil
-}
-
-func requireStringArray(parent map[string]any, key string, path string) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		return fmt.Errorf("%s.%s is required", path, key)
-	}
-	items, ok := value.([]any)
-	if !ok {
-		return fmt.Errorf("%s.%s must be an array", path, key)
-	}
-	for index, item := range items {
-		if _, ok := item.(string); !ok {
-			return fmt.Errorf("%s.%s[%d] must be a string", path, key, index)
-		}
-	}
-	return nil
-}
-
-func requireNumber(parent map[string]any, key string, path string) error {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		return fmt.Errorf("%s.%s is required", path, key)
-	}
-	if _, ok := value.(float64); !ok {
-		return fmt.Errorf("%s.%s must be a number", path, key)
-	}
-	return nil
-}
-
-func factoryLayoutAPIFromInternal(layout *interfaces.FactoryLayoutConfig) *factoryapi.FactoryLayout {
-	if layout == nil {
-		return nil
-	}
-
-	apiLayout := &factoryapi.FactoryLayout{
-		SchemaVersion: int32(layout.SchemaVersion),
-		Nodes:         factoryLayoutNodesAPIFromInternal(layout.Nodes),
-		Edges:         factoryLayoutEdgesAPIFromInternal(layout.Edges),
-		Groups:        factoryLayoutGroupsAPIFromInternal(layout.Groups),
-		Annotations:   factoryLayoutAnnotationsAPIFromInternal(layout.Annotations),
-		Viewport:      factoryLayoutViewportAPIFromInternal(layout.Viewport),
-		Preferences:   factoryLayoutPreferencesAPIFromInternal(layout.Preferences),
-	}
-	return apiLayout
-}
-
-func factoryLayoutNodesAPIFromInternal(nodes []interfaces.FactoryLayoutNodeConfig) *[]factoryapi.FactoryLayoutNode {
-	if len(nodes) == 0 {
-		return nil
-	}
-	values := make([]factoryapi.FactoryLayoutNode, len(nodes))
-	for i, node := range nodes {
-		values[i] = factoryapi.FactoryLayoutNode{
-			Id:         node.ID,
-			Position:   factoryLayoutPointAPIFromInternal(node.Position),
-			Size:       factoryLayoutSizeAPIFromInternal(node.Size),
-			Locked:     node.Locked,
-			EmptyState: factoryLayoutEmptyStateAPIFromInternal(node.EmptyState),
-		}
-	}
-	return &values
-}
-
-func factoryLayoutEdgesAPIFromInternal(edges []interfaces.FactoryLayoutEdgeConfig) *[]factoryapi.FactoryLayoutEdge {
-	if len(edges) == 0 {
-		return nil
-	}
-	values := make([]factoryapi.FactoryLayoutEdge, len(edges))
-	for i, edge := range edges {
-		values[i] = factoryapi.FactoryLayoutEdge{
-			Id:            edge.ID,
-			Waypoints:     factoryLayoutPointsAPIFromInternal(edge.Waypoints),
-			LabelPosition: factoryLayoutPointPtrAPIFromInternal(edge.LabelPosition),
-		}
-	}
-	return &values
-}
-
-func factoryLayoutGroupsAPIFromInternal(groups []interfaces.FactoryLayoutGroupConfig) *[]factoryapi.FactoryLayoutGroup {
-	if len(groups) == 0 {
-		return nil
-	}
-	values := make([]factoryapi.FactoryLayoutGroup, len(groups))
-	for i, group := range groups {
-		values[i] = factoryapi.FactoryLayoutGroup{
-			Id:            group.ID,
-			Label:         stringPtrIfNotEmpty(group.Label),
-			Bounds:        factoryLayoutBoundsAPIFromInternal(group.Bounds),
-			NodeIds:       append([]string(nil), group.NodeIDs...),
-			ParentGroupId: group.ParentGroupID,
-			Color:         stringPtrIfNotEmpty(group.Color),
-			Locked:        group.Locked,
-		}
-	}
-	return &values
-}
-
-func factoryLayoutViewportAPIFromInternal(viewport *interfaces.FactoryLayoutViewportConfig) *factoryapi.FactoryLayoutViewport {
-	if viewport == nil {
-		return nil
-	}
-	return &factoryapi.FactoryLayoutViewport{
-		X:    float32(viewport.X),
-		Y:    float32(viewport.Y),
-		Zoom: float32(viewport.Zoom),
-	}
-}
-
-func factoryLayoutPreferencesAPIFromInternal(preferences *interfaces.FactoryLayoutPreferencesConfig) *factoryapi.FactoryLayoutPreferences {
-	if preferences == nil {
-		return nil
-	}
-	result := &factoryapi.FactoryLayoutPreferences{}
-	if strings.TrimSpace(preferences.Direction) != "" {
-		direction := factoryapi.FactoryLayoutPreferencesDirection(preferences.Direction)
-		result.Direction = &direction
-	}
-	return result
-}
-
-func factoryLayoutPointAPIFromInternal(point interfaces.FactoryLayoutPointConfig) factoryapi.FactoryLayoutPoint {
-	return factoryapi.FactoryLayoutPoint{
-		X: float32(point.X),
-		Y: float32(point.Y),
-	}
-}
-
-func factoryLayoutPointPtrAPIFromInternal(point *interfaces.FactoryLayoutPointConfig) *factoryapi.FactoryLayoutPoint {
-	if point == nil {
-		return nil
-	}
-	value := factoryLayoutPointAPIFromInternal(*point)
-	return &value
-}
-
-func factoryLayoutPointsAPIFromInternal(points []interfaces.FactoryLayoutPointConfig) *[]factoryapi.FactoryLayoutPoint {
-	if len(points) == 0 {
-		return nil
-	}
-	values := make([]factoryapi.FactoryLayoutPoint, len(points))
-	for i, point := range points {
-		values[i] = factoryLayoutPointAPIFromInternal(point)
-	}
-	return &values
-}
-
-func factoryLayoutSizeAPIFromInternal(size *interfaces.FactoryLayoutSizeConfig) *factoryapi.FactoryLayoutSize {
-	if size == nil {
-		return nil
-	}
-	return &factoryapi.FactoryLayoutSize{
-		Width:  float32(size.Width),
-		Height: float32(size.Height),
-	}
-}
-
-func factoryLayoutBoundsAPIFromInternal(bounds interfaces.FactoryLayoutBoundsConfig) factoryapi.FactoryLayoutBounds {
-	return factoryapi.FactoryLayoutBounds{
-		X:      float32(bounds.X),
-		Y:      float32(bounds.Y),
-		Width:  float32(bounds.Width),
-		Height: float32(bounds.Height),
-	}
-}
 
 func inputTypesAPIFromInternal(inputTypes []interfaces.InputTypeConfig) *[]factoryapi.InputType {
 	if len(inputTypes) == 0 {
@@ -898,6 +37,7 @@ func invocationReturnAPIFromInternal(value *interfaces.InvocationReturnConfig) *
 	}
 	if strings.TrimSpace(value.WorkTypeName) != "" {
 		result.WorkTypeName = stringPtrIfNotEmpty(value.WorkTypeName)
+
 	}
 	if strings.TrimSpace(value.TerminalState) != "" {
 		result.TerminalState = stringPtrIfNotEmpty(value.TerminalState)
@@ -1347,7 +487,8 @@ func modelOperationSlotsAPIFromInternal(slots []interfaces.ModelOperationSlot) *
 	values := make([]factoryapi.ModelOperationSlot, len(slots))
 	for i, slot := range slots {
 		values[i] = factoryapi.ModelOperationSlot{
-			Name:         slot.Name,
+			Name: slot.Name,
+
 			ContentTypes: modelOperationContentTypesAPIFromInternal(slot.ContentTypes),
 			Required:     boolPtrIfTrue(slot.Required),
 		}
