+++ pkg/transports/mapping/factoryconfig/factory_config_mapping_internal.go
@@ -1,9 +1,7 @@
-// backendsizecheck:ignore-file canonical public factory contract internal mapping remains consolidated until dedicated mapper seams are extracted from pkg/config.
-// pkgmaintcheck:ignore-file-lines canonical public factory contract internal mapping remains consolidated until dedicated mapper seams are extracted from pkg/config.
+// Factory mapping helpers for the remaining work-type, resource, worker, workstation, guard, and normalization projections.
 package factoryconfig
 
 import (
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"strings"
@@ -12,300 +10,6 @@ import (
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	contentcontract "github.com/portpowered/infinite-you/pkg/transports/mapping/workcontent"
 )
-
-func factoryInternalFromAPI(apiCfg factoryapi.Factory) (interfaces.FactoryConfig, error) {
-	cfg := interfaces.FactoryConfig{Name: string(apiCfg.Name)}
-	if err := mapFactoryMetadataInternalFromAPI(&cfg, apiCfg.Description, apiCfg.Examples); err != nil {
-		return interfaces.FactoryConfig{}, err
-	}
-	if apiCfg.Id != nil {
-		cfg.Project = *apiCfg.Id
-	}
-	if apiCfg.Version != nil {
-		cfg.Version = &interfaces.FactoryVersion{
-			Logical:  apiCfg.Version.Logical.Int64(),
-			Physical: apiCfg.Version.Physical.UTC(),
-		}
-	}
-	cfg.Guards = factoryGuardsInternalFromAPI(apiCfg.Guards)
-	if apiCfg.InputTypes != nil {
-		cfg.InputTypes = inputTypesInternalFromAPI(*apiCfg.InputTypes)
-	}
-	cfg.InvocationReturn = invocationReturnInternalFromAPI(apiCfg.InvocationReturn)
-	cfg.InvocationSignature = invocationSignatureInternalFromAPI(apiCfg.InvocationSignature)
-	cfg.Webhooks = factoryWebhooksInternalFromAPI(apiCfg.Webhooks)
-	orchestrator, err := orchestratorInternalFromAPI(apiCfg.Orchestrator)
-	if err != nil {
-		return interfaces.FactoryConfig{}, err
-	}
-	cfg.Orchestrator = orchestrator
-	if apiCfg.WorkTypes != nil {
-		workTypes, err := workTypesInternalFromAPI(*apiCfg.WorkTypes)
-		if err != nil {
-			return interfaces.FactoryConfig{}, err
-		}
-		cfg.WorkTypes = workTypes
-	}
-	if apiCfg.Resources != nil {
-		cfg.Resources = resourcesInternalFromAPI(*apiCfg.Resources)
-	}
-	if apiCfg.SupportingFiles != nil {
-		cfg.ResourceManifest = resourceManifestInternalFromAPI(apiCfg.SupportingFiles)
-	}
-	if apiCfg.Layout != nil {
-		cfg.Layout = factoryLayoutInternalFromAPI(apiCfg.Layout)
-	}
-	if apiCfg.Workers != nil {
-		workers, err := workersInternalFromAPI(*apiCfg.Workers)
-		if err != nil {
-			return interfaces.FactoryConfig{}, err
-		}
-		cfg.Workers = workers
-	}
-	if apiCfg.Workstations != nil {
-		workstations, err := workstationsInternalFromAPI(*apiCfg.Workstations)
-		if err != nil {
-			return interfaces.FactoryConfig{}, err
-		}
-		cfg.Workstations = workstations
-	}
-	return cfg, nil
-}
-
-func factoryWebhooksInternalFromAPI(values *[]factoryapi.FactoryWebhook) []interfaces.FactoryWebhookConfig {
-	if values == nil {
-		return nil
-	}
-	result := make([]interfaces.FactoryWebhookConfig, len(*values))
-	for index, value := range *values {
-		webhook := interfaces.FactoryWebhookConfig{
-			Name:             value.Name,
-			Enabled:          value.Enabled,
-			URL:              value.Url,
-			SigningSecretRef: value.SigningSecretRef,
-			Filter: interfaces.FactoryWebhookFilterConfig{
-				EventTypes: make([]string, len(value.Filter.EventTypes)),
-			},
-			DeliveryPolicy: factoryWebhookDeliveryPolicyInternalFromAPI(value.DeliveryPolicy),
-		}
-		for eventIndex, eventType := range value.Filter.EventTypes {
-			webhook.Filter.EventTypes[eventIndex] = string(eventType)
-		}
-		if value.Filter.DispatchStatuses != nil {
-			webhook.Filter.DispatchStatuses = make([]string, len(*value.Filter.DispatchStatuses))
-			for statusIndex, status := range *value.Filter.DispatchStatuses {
-				webhook.Filter.DispatchStatuses[statusIndex] = string(status)
-			}
-		}
-		result[index] = webhook
-	}
-	return result
-}
-
-func factoryWebhookDeliveryPolicyInternalFromAPI(value *factoryapi.FactoryWebhookDeliveryPolicy) *interfaces.FactoryWebhookDeliveryPolicyConfig {
-	if value == nil {
-		return nil
-	}
-	result := &interfaces.FactoryWebhookDeliveryPolicyConfig{
-		RequestTimeout: value.RequestTimeout,
-		InitialBackoff: value.InitialBackoff,
-		MaxBackoff:     value.MaxBackoff,
-	}
-	if value.MaxAttempts != nil {
-		attempts := *value.MaxAttempts
-		result.MaxAttempts = &attempts
-	}
-	if value.BackoffMultiplier != nil {
-		multiplier := float64(*value.BackoffMultiplier)
-		result.BackoffMultiplier = &multiplier
-	}
-	return result
-}
-
-func mapFactoryMetadataInternalFromAPI(cfg *interfaces.FactoryConfig, description *factoryapi.NameValue, examples *[]factoryapi.FactoryInvocationExample) error {
-	mappedDescription, err := nameValueInternalFromAPI(description, "factory.description")
-	if err != nil {
-		return err
-	}
-	mappedExamples, err := invocationExamplesInternalFromAPI(examples)
-	if err != nil {
-		return err
-	}
-	cfg.Description = mappedDescription
-	cfg.Examples = mappedExamples
-	return nil
-}
-
-// FactoryConfigFromOpenAPI converts the generated OpenAPI factory model into
-// the internal config representation.
-func FactoryConfigFromOpenAPI(apiCfg factoryapi.Factory) (interfaces.FactoryConfig, error) {
-	return factoryInternalFromAPI(apiCfg)
-}
-
-// NameValueFromOpenAPI validates and maps one generated localized value into
-// the transport-independent Factory definition contract.
-func NameValueFromOpenAPI(value factoryapi.NameValue) (interfaces.NameValueConfig, error) {
-	return nameValueFromOpenAPI(value, "")
-}
-
-func nameValueInternalFromAPI(value *factoryapi.NameValue, fieldPath string) (*interfaces.NameValueConfig, error) {
-	if value == nil {
-		return nil, nil
-	}
-	mapped, err := nameValueFromOpenAPI(*value, fieldPath)
-	if err != nil {
-		return nil, err
-	}
-	return &mapped, nil
-}
-
-func nameValueFromOpenAPI(value factoryapi.NameValue, fieldPath string) (interfaces.NameValueConfig, error) {
-	result := interfaces.NameValueConfig{
-		Type:  string(value.Type),
-		Value: value.Value,
-		ID:    stringValue(value.Id),
-	}
-	if value.Locales != nil {
-		result.Locales = append([]string(nil), (*value.Locales)...)
-	}
-	if value.Values != nil {
-		result.Values = make(map[string]string, len(*value.Values))
-		for locale, localized := range *value.Values {
-			result.Values[locale] = localized
-		}
-	}
-	if err := result.Validate(); err != nil {
-		if fieldPath != "" {
-			return interfaces.NameValueConfig{}, fmt.Errorf("%s.%w", fieldPath, err)
-		}
-		return interfaces.NameValueConfig{}, err
-	}
-	return result, nil
-}
-
-func inputTypesInternalFromAPI(inputTypes []factoryapi.InputType) []interfaces.InputTypeConfig {
-	values := make([]interfaces.InputTypeConfig, len(inputTypes))
-	for i, inputType := range inputTypes {
-		values[i] = interfaces.InputTypeConfig{
-			Name: inputType.Name,
-			Type: internalFactoryInputKindFromPublic(inputType.Type),
-		}
-	}
-	return values
-}
-
-func invocationReturnInternalFromAPI(value *factoryapi.InvocationReturn) *interfaces.InvocationReturnConfig {
-	if value == nil {
-		return nil
-	}
-	return &interfaces.InvocationReturnConfig{
-		Policy:        string(value.Policy),
-		WorkTypeName:  stringValue(value.WorkTypeName),
-		TerminalState: stringValue(value.TerminalState),
-		WorkName:      stringValue(value.WorkName),
-	}
-}
-
-func invocationSignatureInternalFromAPI(value *factoryapi.FactoryInvocationSignature) *interfaces.InvocationSignatureConfig {
-	if value == nil {
-		return nil
-	}
-	return &interfaces.InvocationSignatureConfig{
-		Parameters:                 invocationParametersInternalFromAPI(value.Parameters),
-		UnknownNamedArgumentPolicy: enumStringValue(value.UnknownNamedArgumentPolicy),
-		OutputContract:             invocationOutputContractInternalFromAPI(value.OutputContract),
-	}
-}
-
-func invocationParametersInternalFromAPI(parameters *[]factoryapi.FactoryInvocationParameter) []interfaces.InvocationParameterConfig {
-	if parameters == nil {
-		return nil
-	}
-	values := make([]interfaces.InvocationParameterConfig, len(*parameters))
-	for i, parameter := range *parameters {
-		values[i] = interfaces.InvocationParameterConfig{
-			Name:          parameter.Name,
-			Description:   stringValue(parameter.Description),
-			ExternalName:  stringValue(parameter.ExternalName),
-			Aliases:       stringSliceValue(parameter.Aliases),
-			TypeHint:      enumStringValue(parameter.TypeHint),
-			ValueMode:     enumStringValue(parameter.ValueMode),
-			Required:      boolValue(parameter.Required),
-			Sensitive:     boolValue(parameter.Sensitive),
-			Choices:       stringSliceValue(parameter.Choices),
-			DefaultValue:  stringValue(parameter.DefaultValue),
-			DefaultValues: stringSliceValue(parameter.DefaultValues),
-			Bindings:      invocationParameterBindingsInternalFromAPI(parameter.Bindings),
-		}
-	}
-	return values
-}
-
-func invocationParameterBindingsInternalFromAPI(bindings *[]factoryapi.FactoryInvocationParameterBinding) []interfaces.InvocationParameterBindingConfig {
-	if bindings == nil {
-		return nil
-	}
-	values := make([]interfaces.InvocationParameterBindingConfig, len(*bindings))
-	for i, binding := range *bindings {
-		values[i] = interfaces.InvocationParameterBindingConfig{
-			Kind:     string(binding.Kind),
-			Position: intValue(binding.Position),
-		}
-	}
-	return values
-}
-
-func invocationOutputContractInternalFromAPI(value *factoryapi.FactoryInvocationOutputContract) *interfaces.InvocationOutputContractConfig {
-	if value == nil {
-		return nil
-	}
-	return &interfaces.InvocationOutputContractConfig{
-		Mode:          enumStringValue(value.Mode),
-		PathParameter: stringValue(value.PathParameter),
-		ContentType:   stringValue(value.ContentType),
-		FileExtension: stringValue(value.FileExtension),
-		Description:   stringValue(value.Description),
-	}
-}
-
-func invocationExamplesInternalFromAPI(examples *[]factoryapi.FactoryInvocationExample) ([]interfaces.InvocationExampleConfig, error) {
-	if examples == nil {
-		return nil, nil
-	}
-	values := make([]interfaces.InvocationExampleConfig, len(*examples))
-	for i, example := range *examples {
-		description, err := nameValueFromOpenAPI(example.Description, fmt.Sprintf("factory.examples[%d].description", i))
-		if err != nil {
-			return nil, err
-		}
-		args, err := invocationExampleArgsInternalFromAPI(example.Args)
-		if err != nil {
-			return nil, fmt.Errorf("factory.examples[%d].args: %w", i, err)
-		}
-		values[i] = interfaces.InvocationExampleConfig{
-			Name:        example.Name,
-			Description: description,
-			Args:        args,
-		}
-	}
-	return values, nil
-}
-
-func invocationExampleArgsInternalFromAPI(args factoryapi.FactoryInvocationArguments) (interfaces.InvocationExampleArguments, error) {
-	values := make(interfaces.InvocationExampleArguments, len(args))
-	for name, value := range args {
-		if scalar, err := value.AsFactoryInvocationArguments0(); err == nil {
-			values[name] = scalar
-			continue
-		}
-		list, err := value.AsFactoryInvocationArguments1()
-		if err != nil {
-			return nil, fmt.Errorf("%s must be a string or array of strings", name)
-		}
-		values[name] = append([]string(nil), list...)
-	}
-	return values, nil
-}
 
 func workTypeHandlingBehaviorInternalFromAPI(behaviors *[]factoryapi.WorkTypeHandlingBehavior) []string {
 	if behaviors == nil || len(*behaviors) == 0 {
@@ -382,201 +86,6 @@ func resourceManifestInternalFromAPI(manifest *factoryapi.ResourceManifest) *int
 		return &interfaces.PortableResourceManifestConfig{}
 	}
 	return cfg
-}
-
-func factoryLayoutInternalFromAPI(layout *factoryapi.FactoryLayout) *interfaces.FactoryLayoutConfig {
-	if layout == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutConfig{
-		SchemaVersion: int(layout.SchemaVersion),
-		Nodes:         factoryLayoutNodesInternalFromAPI(layout.Nodes),
-		Edges:         factoryLayoutEdgesInternalFromAPI(layout.Edges),
-		Groups:        factoryLayoutGroupsInternalFromAPI(layout.Groups),
-		Annotations:   factoryLayoutAnnotationsInternalFromAPI(layout.Annotations),
-		Viewport:      factoryLayoutViewportInternalFromAPI(layout.Viewport),
-		Preferences:   factoryLayoutPreferencesInternalFromAPI(layout.Preferences),
-	}
-}
-
-func factoryLayoutAnnotationsInternalFromAPI(annotations *[]factoryapi.FactoryLayoutAnnotation) []interfaces.FactoryLayoutAnnotationConfig {
-	if annotations == nil {
-		return nil
-	}
-	values := make([]interfaces.FactoryLayoutAnnotationConfig, len(*annotations))
-	for i, annotation := range *annotations {
-		values[i] = interfaces.FactoryLayoutAnnotationConfig{
-			ID:       annotation.Id,
-			Kind:     string(annotation.Kind),
-			Position: factoryLayoutAnnotationPositionInternalFromAPI(annotation.Position),
-			Size:     factoryLayoutAnnotationSizeInternalFromAPI(annotation.Size),
-			Note:     factoryLayoutNoteInternalFromAPI(annotation.Note),
-			Image:    factoryLayoutImageInternalFromAPI(annotation.Image),
-		}
-	}
-	return values
-}
-
-func factoryLayoutAnnotationPositionInternalFromAPI(position factoryapi.FactoryLayoutAnnotationPosition) interfaces.FactoryLayoutPointConfig {
-	return interfaces.FactoryLayoutPointConfig{X: float64(position.X), Y: float64(position.Y)}
-}
-
-func factoryLayoutAnnotationSizeInternalFromAPI(size *factoryapi.FactoryLayoutAnnotationSize) *interfaces.FactoryLayoutSizeConfig {
-	if size == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutSizeConfig{Width: float64(size.Width), Height: float64(size.Height)}
-}
-
-func factoryLayoutNoteInternalFromAPI(note *factoryapi.FactoryLayoutNote) *interfaces.FactoryLayoutNoteConfig {
-	if note == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutNoteConfig{
-		Title: stringValue(note.Title),
-		Body:  note.Body,
-		Tone:  string(note.Tone),
-	}
-}
-
-func factoryLayoutImageInternalFromAPI(image *factoryapi.FactoryLayoutImage) *interfaces.FactoryLayoutImageConfig {
-	if image == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutImageConfig{
-		Source: interfaces.FactoryLayoutImageSourceConfig{
-			Kind:      string(image.Source.Kind),
-			MediaType: string(image.Source.MediaType),
-			Data:      base64.StdEncoding.EncodeToString(image.Source.Data),
-		},
-		AlternativeText: image.AlternativeText,
-	}
-}
-
-func factoryLayoutNodesInternalFromAPI(nodes *[]factoryapi.FactoryLayoutNode) []interfaces.FactoryLayoutNodeConfig {
-	if nodes == nil {
-		return nil
-	}
-	values := make([]interfaces.FactoryLayoutNodeConfig, len(*nodes))
-	for i, node := range *nodes {
-		values[i] = interfaces.FactoryLayoutNodeConfig{
-			ID:         node.Id,
-			Position:   factoryLayoutPointInternalFromAPI(node.Position),
-			Size:       factoryLayoutSizeInternalFromAPI(node.Size),
-			Locked:     node.Locked,
-			EmptyState: factoryLayoutEmptyStateInternalFromAPI(node.EmptyState),
-		}
-	}
-	return values
-}
-
-func factoryLayoutEmptyStateInternalFromAPI(emptyState *factoryapi.FactoryLayoutEmptyState) *interfaces.FactoryLayoutEmptyStateConfig {
-	if emptyState == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutEmptyStateConfig{
-		Text:  stringValue(emptyState.Text),
-		Image: factoryLayoutImageInternalFromAPI(emptyState.Image),
-	}
-}
-
-func factoryLayoutEdgesInternalFromAPI(edges *[]factoryapi.FactoryLayoutEdge) []interfaces.FactoryLayoutEdgeConfig {
-	if edges == nil {
-		return nil
-	}
-	values := make([]interfaces.FactoryLayoutEdgeConfig, len(*edges))
-	for i, edge := range *edges {
-		values[i] = interfaces.FactoryLayoutEdgeConfig{
-			ID:            edge.Id,
-			Waypoints:     factoryLayoutPointsInternalFromAPI(edge.Waypoints),
-			LabelPosition: factoryLayoutPointPtrInternalFromAPI(edge.LabelPosition),
-		}
-	}
-	return values
-}
-
-func factoryLayoutGroupsInternalFromAPI(groups *[]factoryapi.FactoryLayoutGroup) []interfaces.FactoryLayoutGroupConfig {
-	if groups == nil {
-		return nil
-	}
-	values := make([]interfaces.FactoryLayoutGroupConfig, len(*groups))
-	for i, group := range *groups {
-		values[i] = interfaces.FactoryLayoutGroupConfig{
-			ID:            group.Id,
-			Label:         stringValue(group.Label),
-			Bounds:        factoryLayoutBoundsInternalFromAPI(group.Bounds),
-			NodeIDs:       append([]string(nil), group.NodeIds...),
-			ParentGroupID: group.ParentGroupId,
-			Color:         stringValue(group.Color),
-			Locked:        group.Locked,
-		}
-	}
-	return values
-}
-
-func factoryLayoutViewportInternalFromAPI(viewport *factoryapi.FactoryLayoutViewport) *interfaces.FactoryLayoutViewportConfig {
-	if viewport == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutViewportConfig{
-		X:    float64(viewport.X),
-		Y:    float64(viewport.Y),
-		Zoom: float64(viewport.Zoom),
-	}
-}
-
-func factoryLayoutPreferencesInternalFromAPI(preferences *factoryapi.FactoryLayoutPreferences) *interfaces.FactoryLayoutPreferencesConfig {
-	if preferences == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutPreferencesConfig{
-		Direction: enumStringValue(preferences.Direction),
-	}
-}
-
-func factoryLayoutPointInternalFromAPI(point factoryapi.FactoryLayoutPoint) interfaces.FactoryLayoutPointConfig {
-	return interfaces.FactoryLayoutPointConfig{
-		X: float64(point.X),
-		Y: float64(point.Y),
-	}
-}
-
-func factoryLayoutPointPtrInternalFromAPI(point *factoryapi.FactoryLayoutPoint) *interfaces.FactoryLayoutPointConfig {
-	if point == nil {
-		return nil
-	}
-	value := factoryLayoutPointInternalFromAPI(*point)
-	return &value
-}
-
-func factoryLayoutPointsInternalFromAPI(points *[]factoryapi.FactoryLayoutPoint) []interfaces.FactoryLayoutPointConfig {
-	if points == nil {
-		return nil
-	}
-	values := make([]interfaces.FactoryLayoutPointConfig, len(*points))
-	for i, point := range *points {
-		values[i] = factoryLayoutPointInternalFromAPI(point)
-	}
-	return values
-}
-
-func factoryLayoutSizeInternalFromAPI(size *factoryapi.FactoryLayoutSize) *interfaces.FactoryLayoutSizeConfig {
-	if size == nil {
-		return nil
-	}
-	return &interfaces.FactoryLayoutSizeConfig{
-		Width:  float64(size.Width),
-		Height: float64(size.Height),
-	}
-}
-
-func factoryLayoutBoundsInternalFromAPI(bounds factoryapi.FactoryLayoutBounds) interfaces.FactoryLayoutBoundsConfig {
-	return interfaces.FactoryLayoutBoundsConfig{
-		X:      float64(bounds.X),
-		Y:      float64(bounds.Y),
-		Width:  float64(bounds.Width),
-		Height: float64(bounds.Height),
-	}
 }
 
 func requiredToolsInternalFromAPI(requiredTools *[]factoryapi.RequiredTool) []interfaces.RequiredToolConfig {
@@ -898,8 +407,9 @@ func workstationOperationBindingSelectorInternalFromAPI(selector *factoryapi.Wor
 	return &interfaces.ModelOperationBindingSelector{
 		Slot:  stringValue(selector.Slot),
 		Label: stringValue(selector.Label),
-		Type:  internalFactoryModelOperationContentTypeFromPublic(valueOrEmpty(selector.Type)),
-		Role:  stringValue(selector.Role),
+
+		Type: internalFactoryModelOperationContentTypeFromPublic(valueOrEmpty(selector.Type)),
+		Role: stringValue(selector.Role),
 	}
 }
 
@@ -1348,6 +858,7 @@ func internalFactoryGuardTypeFromPublic(value factoryapi.GuardType) interfaces.G
 	case publicFactoryGuardTypeSameTraceID:
 		return interfaces.GuardTypeSameTraceID
 	case publicFactoryGuardTypeInferenceThrottle:
+
 		return interfaces.GuardTypeInferenceThrottle
 	default:
 		return interfaces.GuardType(strings.TrimSpace(string(value)))
@@ -1376,165 +887,6 @@ func publicInputGuardTypeFromInternal(value interfaces.GuardType) factoryapi.Inp
 
 func publicFactoryRootGuardTypeFromInternal(value interfaces.GuardType) factoryapi.FactoryGuardType {
 	return factoryapi.FactoryGuardType(publicFactoryGuardTypeStringFromInternal(value))
-}
-
-func orchestratorInternalFromAPI(value *factoryapi.FactoryOrchestrator) (*interfaces.FactoryOrchestratorConfig, error) {
-	if value == nil {
-		return nil, nil
-	}
-	kind := interfaces.StrictPublicFactoryOrchestratorKind(string(value.Kind))
-	if kind == "" {
-		return &interfaces.FactoryOrchestratorConfig{Kind: string(value.Kind)}, nil
-	}
-	cfg := &interfaces.FactoryOrchestratorConfig{
-		Kind: kind,
-	}
-	if value.Petri != nil {
-		cfg.Petri = &interfaces.FactoryOrchestratorPetriConfig{}
-	}
-	if value.Javascript != nil {
-		jsCfg, err := orchestratorJavaScriptInternalFromAPI(*value.Javascript)
-		if err != nil {
-			return nil, err
-		}
-		cfg.JavaScript = jsCfg
-	}
-	return cfg, nil
-}
-
-func orchestratorJavaScriptInternalFromAPI(value factoryapi.FactoryOrchestratorJavaScriptConfig) (*interfaces.FactoryOrchestratorJavaScriptConfig, error) {
-	cfg := &interfaces.FactoryOrchestratorJavaScriptConfig{
-		Dialect:    stringValue(value.Dialect),
-		SourceRef:  stringValue(value.SourceRef),
-		SourceHash: stringValue(value.SourceHash),
-		Entrypoint: stringValue(value.Entrypoint),
-	}
-	if value.Metadata != nil {
-		cfg.Metadata = map[string]string(*value.Metadata)
-	}
-	if value.InlineSource != nil {
-		cfg.InlineSource = &interfaces.FactoryOrchestratorJavaScriptInlineSource{
-			Encoding: string(value.InlineSource.Encoding),
-			Inline:   value.InlineSource.Inline,
-		}
-	}
-	if value.ArgsSchema != nil {
-		raw, err := json.Marshal(value.ArgsSchema)
-		if err != nil {
-			return nil, err
-		}
-		cfg.ArgsSchema = raw
-	}
-	if value.DefaultPolicy != nil {
-		raw, err := json.Marshal(value.DefaultPolicy)
-		if err != nil {
-			return nil, err
-		}
-		cfg.DefaultPolicy = raw
-	}
-	if value.Agents != nil {
-		cfg.Agents = make(map[string]interfaces.FactoryOrchestratorJavaScriptAgent, len(*value.Agents))
-		for id, agent := range *value.Agents {
-			cfg.Agents[id] = interfaces.FactoryOrchestratorJavaScriptAgent{Preset: agent.Preset}
-		}
-	}
-	return cfg, nil
-}
-
-func orchestratorAPIFromInternal(cfg *interfaces.FactoryConfig) *factoryapi.FactoryOrchestrator {
-	if cfg == nil || cfg.Orchestrator == nil {
-		return nil
-	}
-	kind := interfaces.EffectiveOrchestratorKind(cfg)
-	apiKind := factoryapi.FactoryOrchestratorKind(interfaces.StrictPublicFactoryOrchestratorKind(kind))
-	result := &factoryapi.FactoryOrchestrator{
-		Kind: apiKind,
-	}
-	if cfg.Orchestrator.Petri != nil || kind == interfaces.OrchestratorKindPetri {
-		result.Petri = &factoryapi.FactoryOrchestratorPetriConfig{}
-	}
-	if cfg.Orchestrator.JavaScript != nil {
-		result.Javascript = orchestratorJavaScriptAPIFromInternal(cfg.Orchestrator.JavaScript)
-	}
-	return result
-}
-
-// ProjectEffectiveOrchestratorForAPIRead fills the compatibility PETRI orchestrator
-// projection when a factory has no authored orchestrator block.
-func ProjectEffectiveOrchestratorForAPIRead(api factoryapi.Factory, cfg *interfaces.FactoryConfig) factoryapi.Factory {
-	if api.Orchestrator != nil {
-		return api
-	}
-	if interfaces.EffectiveOrchestratorKind(cfg) == interfaces.OrchestratorKindPetri {
-		api.Orchestrator = defaultPetriOrchestratorAPI()
-	}
-	return api
-}
-
-func defaultPetriOrchestratorAPI() *factoryapi.FactoryOrchestrator {
-	kind := factoryapi.PETRI
-	return &factoryapi.FactoryOrchestrator{
-		Kind:  kind,
-		Petri: &factoryapi.FactoryOrchestratorPetriConfig{},
-	}
-}
-
-func orchestratorJavaScriptAPIFromInternal(cfg *interfaces.FactoryOrchestratorJavaScriptConfig) *factoryapi.FactoryOrchestratorJavaScriptConfig {
-	if cfg == nil {
-		return nil
-	}
-	result := &factoryapi.FactoryOrchestratorJavaScriptConfig{
-		Dialect:    stringPtrIfNotEmpty(cfg.Dialect),
-		SourceRef:  stringPtrIfNotEmpty(cfg.SourceRef),
-		SourceHash: stringPtrIfNotEmpty(cfg.SourceHash),
-		Entrypoint: stringPtrIfNotEmpty(cfg.Entrypoint),
-	}
-	if len(cfg.Metadata) > 0 {
-		metadata := factoryapi.StringMap(cfg.Metadata)
-		result.Metadata = &metadata
-	}
-	if cfg.InlineSource != nil {
-		result.InlineSource = &factoryapi.FactoryOrchestratorJavaScriptInlineSource{
-			Encoding: factoryapi.FactoryOrchestratorJavaScriptInlineSourceEncoding(cfg.InlineSource.Encoding),
-			Inline:   cfg.InlineSource.Inline,
-		}
-	}
-	if len(cfg.ArgsSchema) > 0 {
-		var argsSchema map[string]any
-		if err := json.Unmarshal(cfg.ArgsSchema, &argsSchema); err == nil {
-			result.ArgsSchema = &argsSchema
-		}
-	}
-	if len(cfg.DefaultPolicy) > 0 {
-		var defaultPolicy map[string]any
-		if err := json.Unmarshal(cfg.DefaultPolicy, &defaultPolicy); err == nil {
-			result.DefaultPolicy = &defaultPolicy
-		}
-	}
-	if len(cfg.Agents) > 0 {
-		agents := make(map[string]factoryapi.FactoryOrchestratorJavaScriptAgent, len(cfg.Agents))
-		for id, agent := range cfg.Agents {
-			agents[id] = factoryapi.FactoryOrchestratorJavaScriptAgent{Preset: agent.Preset}
-		}
-		result.Agents = &agents
-	}
-	return result
-}
-
-func isDefaultPetriOrchestratorAPI(value *factoryapi.FactoryOrchestrator) bool {
-	if value == nil {
-		return true
-	}
-	if value.Kind != factoryapi.PETRI {
-		return false
-	}
-	if value.Javascript != nil {
-		return false
-	}
-	if value.Petri == nil {
-		return true
-	}
-	return true
 }
 
 func cloneFactoryOrchestratorConfig(cfg *interfaces.FactoryOrchestratorConfig) *interfaces.FactoryOrchestratorConfig {
